+++ ios/flutter_bugfender/Package.swift
@@ -13,7 +13,7 @@ let package = Package(
     ],
     dependencies: [
         .package(name: "FlutterFramework", path: "../FlutterFramework"),
-        .package(url: "https://github.com/bugfender/BugfenderSDK-iOS", from: "2.2.0")
+        .package(url: "https://github.com/bugfender/BugfenderSDK-iOS", from: "3.0.0")
     ],
     targets: [
         .target(